+++ ltk_data/lookup.py
@@ -1,0 +1,14 @@
+# =============================================================================
+# C O P Y R I G H T
+# -----------------------------------------------------------------------------
+# Copyright (c) 2021-2022 by Helmut Konrad Fahrendholz. All rights reserved.
+# This file is property of Helmut Konrad Fahrendholz. Any unauthorized copy,
+# use or distribution is an offensive act against international law and may
+# be prosecuted under federal law. Its content is company confidential.
+# =============================================================================
+
+import ltk_data
+
+NAME_FAMILY = ltk_data.load_pickle(ltk_data.NAMES_FAMILY)
+NAME_FEMALE = ltk_data.load_pickle(ltk_data.NAMES_FEMALE)
+NAME_MALE = ltk_data.load_pickle(ltk_data.NAMES_MALE)